Rewrite the client-side keep-alive mechanism
Move most of the logic into a new class: InactivityTimer.
Avoid calling "singleShot"; use a QTimer instance instead with a state
variable.

Hopefully this fixes the bug where a client occasionally loses its
connection to a server on the same machine when waking up the machine
from a sleep state.