Add support for listing the labels currently used
This new feature is only accessible from the command-line remote.

The new command is:

  label list used

The command will output a list of the labels that are considered to be
used. Labels that exist but are no longer associated with any track
will not be included. Labels that are only associated with tracks that
have not been seen in a long time will also not be included.