Player: fix track 'stopped playing' handler
Hopefully this fixes the bug on Linux where the track never seems to
end.

The problem was caused by the mediaStatus not yet being equal to
EndOfMedia at the time when the playbackState changes to StoppedState.
We try to fix this by no longer depending on the mediaStatus when
handling the playbackState change.