Music collection view: add fourth filter field
Add a fourth filter combobox to the music collection view. Any filter
combobox number x will be visible when x-1 or x or any >x is set to
something other than "none". As before, a track needs to satisfy all
filters (not set to "none") in order to be shown in the list.