Add reset button for track filters & highlighting
Add a reset button to each track filter line, and also to the track
highlighting line. Clicking the button will clear the filter or the
highlighting criterium and set it to "none".

For now, this brings back the unwelcome space between the filters and
the highlighting. I will have to find a way to fix that.