Make track filters list size dynamic
In the music collection view, switch from a fixed number of filters (4)
to a list of filters that can grow and shrink. A new filter can be
added at the bottom by clicking the "Add" button. And each filter now
has an extra button (with an icon resembling a trash can) to remove
that filter from the list. The reset buttons are still present; they
set their filter to "none" without removing it, as before.